Move auth form component files' position.
