+++ features/auth/components/form/fields/email.tsx
@@ -1,6 +1,6 @@
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { FormAlert } from "@/features/auth/components/form-alert";
+import { FormAlert } from "@/features/auth/components/form/alert";
 
 function EmailField({ errors }: { errors: string[] | undefined }) {
   return (
